requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
redis==5.0.1
django-redis==5.3.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4